requirement.txt
pytorch==1.9.0
torchvision==0.10.0
matplotlib==3.3.4
scikit-image==0.17.2
timm==0.4.12
kornia==0.5.4
scikit-learn==0.24.2
pandas==1.1.5
psutil==5.8.0